% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{accordion}
\alias{accordion}
\title{Accordion}
\usage{
accordion(..., id = NULL)
}
\arguments{
\item{...}{\code{\link[=accordion_item]{accordion_item()}} elements.}

\item{id}{Optional element ID.}
}
\value{
A Shiny tag object (\verb{<div class="accordion">}).
}
\description{
A container for collapsible \code{\link[=accordion_item]{accordion_item()}} sections.
}
\examples{
accordion(
  accordion_item("Section 1", p("Content one"), open = TRUE),
  accordion_item("Section 2", p("Content two"))
)

}
