% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{card_parts}
\alias{card_parts}
\alias{card_header}
\alias{card_body}
\alias{card_footer}
\title{Card Sub-Components}
\usage{
card_header(...)

card_body(...)

card_footer(...)
}
\arguments{
\item{...}{Content to place inside the element.}
}
\value{
A Shiny tag object.
}
\description{
Helper elements for structuring card content with explicit header, body,
and footer regions.
}
\examples{
div(
  card_header(h3("Title")),
  card_body(p("Body text")),
  card_footer(btn("Action"))
)

}
