% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{divider}
\alias{divider}
\title{Divider Line}
\usage{
divider()
}
\value{
A Shiny tag object (\verb{<hr>}).
}
\description{
A horizontal rule with standard vertical margin, used to separate sections.
}
\examples{
vstack(card("Section 1"), divider(), card("Section 2"))

}
