% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{dropdown_divider}
\alias{dropdown_divider}
\title{Dropdown Divider}
\usage{
dropdown_divider()
}
\value{
A Shiny tag object.
}
\description{
A visual separator line between \code{\link[=dropdown_item]{dropdown_item()}} groups.
}
\examples{
dropdown("Account",
  dropdown_item("Profile"),
  dropdown_divider(),
  dropdown_item("Logout")
)

}
