% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{formatting}
\alias{formatting}
\alias{strong}
\alias{em}
\alias{code}
\alias{pre}
\title{Inline Formatting Elements}
\usage{
strong(...)

em(...)

code(...)

pre(...)
}
\arguments{
\item{...}{Child elements or text content.}
}
\value{
A Shiny tag object.
}
\description{
Wrapper functions for \verb{<strong>}, \verb{<em>}, \verb{<code>}, and \verb{<pre>} tags.
}
\examples{
strong("Bold text")
em("Italic text")
code("x <- 1")
pre("multi\nline")

}
