% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{headings}
\alias{headings}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\title{HTML Heading Elements}
\usage{
h1(..., id = NULL)

h2(..., id = NULL)

h3(..., id = NULL)

h4(..., id = NULL)
}
\arguments{
\item{...}{Child elements or text content passed to the tag.}

\item{id}{Optional element ID.}
}
\value{
A Shiny tag object.
}
\description{
Wrapper functions for \verb{<h1>} through \verb{<h4>} heading tags.
}
\examples{
h1("Page Title")
h2("Section", id = "section-1")

}
