% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{list_group}
\alias{list_group}
\title{List Group}
\usage{
list_group(..., id = NULL)
}
\arguments{
\item{...}{\code{\link[=list_item]{list_item()}} elements.}

\item{id}{Optional element ID.}
}
\value{
A Shiny tag object (\verb{<ul class="list-group">}).
}
\description{
A styled \verb{<ul>} list container. Use with \code{\link[=list_item]{list_item()}}.
}
\examples{
list_group(
  list_item("Apple"),
  list_item("Banana", active = TRUE),
  list_item("Cherry")
)

}
