% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{list_item}
\alias{list_item}
\title{List Group Item}
\usage{
list_item(..., active = FALSE, id = NULL)
}
\arguments{
\item{...}{Content of the list item.}

\item{active}{Logical. Highlight as selected. Default \code{FALSE}.}

\item{id}{Optional element ID.}
}
\value{
A Shiny tag object (\verb{<li class="list-group-item">}).
}
\description{
A single \verb{<li>} item inside a \code{\link[=list_group]{list_group()}}.
}
\examples{
list_item("Normal item")
list_item("Selected item", active = TRUE)

}
