% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{oats_hide_modal}
\alias{oats_hide_modal}
\title{Hide a Modal Dialog}
\usage{
oats_hide_modal(session, id)
}
\arguments{
\item{session}{The Shiny \code{session} object.}

\item{id}{Character. The \code{id} of the \code{\link[=modal]{modal()}} to close.}
}
\value{
Called for its side effect; returns \code{NULL} invisibly.
}
\description{
Programmatically closes a \code{\link[=modal]{modal()}} element defined in the UI.
}
\examples{
if (interactive()) {
  server <- function(input, output, session) {
    observeEvent(input$cancel_btn, {
      oats_hide_modal(session, "my_modal")
    })
  }
}

}
