% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{oats_update_progress}
\alias{oats_update_progress}
\title{Update a Progress Bar Value}
\usage{
oats_update_progress(session, id, value)
}
\arguments{
\item{session}{The Shiny \code{session} object.}

\item{id}{Character. The \code{id} of the \code{\link[=progress_bar]{progress_bar()}}.}

\item{value}{Numeric. The new current value.}
}
\value{
Called for its side effect; returns \code{NULL} invisibly.
}
\description{
Programmatically updates the fill level of a \code{\link[=progress_bar]{progress_bar()}}.
}
\examples{
if (interactive()) {
  server <- function(input, output, session) {
    observe({
      oats_update_progress(session, "upload_progress", value = 75)
    })
  }
}

}
