% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{row}
\alias{row}
\title{Grid Row}
\usage{
row(..., id = NULL)
}
\arguments{
\item{...}{\code{\link[=col]{col()}} elements or other content.}

\item{id}{Optional element ID.}
}
\value{
A Shiny tag object (\verb{<div class="row">}).
}
\description{
A horizontal row that groups \code{\link[=col]{col()}} columns.
}
\examples{
row(
  col(6, p("Left")),
  col(6, p("Right"))
)

}
