% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\docType{package}
\name{shinyoats-package}
\alias{shinyoats}
\alias{shinyoats-package}
\title{shinyoats: Oat UI Components for Shiny}
\description{
A lightweight UI library providing 55 ready-to-use components built on the
Oat UI framework. Cleaner and faster than shinyfluent, with zero external
dependencies beyond Shiny.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pawanramamali/shinyoats}
  \item Report bugs at \url{https://github.com/pawanramamali/shinyoats/issues}
}

}
\author{
\strong{Maintainer}: Pawan Rama Mali \email{prm@outlook.in} (\href{https://orcid.org/0000-0001-7864-5819}{ORCID})

Other contributors:
\itemize{
  \item Kailash Nadh (Oat UI library author) [copyright holder]
}

}
\keyword{internal}
