% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{sidebar_menu}
\alias{sidebar_menu}
\title{Sidebar Menu}
\usage{
sidebar_menu(...)
}
\arguments{
\item{...}{\code{\link[=menu_item]{menu_item()}} elements.}
}
\value{
A Shiny tag object.
}
\description{
A vertical list of \code{\link[=menu_item]{menu_item()}} links inside a \code{\link[=sidebar]{sidebar()}}.
}
\examples{
sidebar_menu(
  menu_item("Home", active = TRUE),
  menu_item("Reports")
)

}
