% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{spacer}
\alias{spacer}
\title{Flex Spacer}
\usage{
spacer()
}
\value{
A Shiny tag object.
}
\description{
A \code{flex: 1} element that pushes siblings to opposite ends of an \code{\link[=hstack]{hstack()}}.
}
\examples{
hstack(span("Left"), spacer(), span("Right"))

}
