% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{spinner}
\alias{spinner}
\title{Loading Spinner}
\usage{
spinner(size = "md", variant = "primary")
}
\arguments{
\item{size}{Character. Spinner size: \code{"md"} (default), \code{"sm"}, or \code{"lg"}.}

\item{variant}{Character. Colour variant. Default \code{"primary"}.}
}
\value{
A Shiny tag object.
}
\description{
An animated circular loading indicator.
}
\examples{
spinner()
spinner(size = "lg", variant = "secondary")

}
