% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyoats.R
\name{text_elements}
\alias{text_elements}
\alias{p}
\alias{span}
\alias{div}
\title{HTML Text Elements}
\usage{
p(..., id = NULL)

span(..., id = NULL)

div(..., id = NULL)
}
\arguments{
\item{...}{Child elements or text content.}

\item{id}{Optional element ID.}
}
\value{
A Shiny tag object.
}
\description{
Wrapper functions for common inline and block text tags.
}
\examples{
p("A paragraph of text.")
span("Inline text", id = "note")
div(h2("Title"), p("Body"))

}
